Move implementation of ank apply into own file
Issue-Id: #229
