+++ ank/src/cli_error.rs
@@ -1,0 +1,48 @@
+use std::fmt;
+
+use crate::cli_commands::server_connection;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum CliError {
+    YamlSerialization(String),
+    JsonSerialization(String),
+    ExecutionError(String),
+}
+
+impl fmt::Display for CliError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            CliError::YamlSerialization(message) => {
+                write!(f, "Could not serialize YAML object: '{message}'")
+            }
+            CliError::JsonSerialization(message) => {
+                write!(f, "Could not serialize JSON object: '{message}'")
+            }
+            CliError::ExecutionError(message) => {
+                write!(f, "Command failed: '{}'", message)
+            }
+        }
+    }
+}
+
+impl From<serde_yaml::Error> for CliError {
+    fn from(value: serde_yaml::Error) -> Self {
+        CliError::YamlSerialization(format!("{value}"))
+    }
+}
+
+impl From<serde_json::Error> for CliError {
+    fn from(value: serde_json::Error) -> Self {
+        CliError::JsonSerialization(format!("{value}"))
+    }
+}
+
+impl From<server_connection::ServerConnectionError> for CliError {
+    fn from(value: server_connection::ServerConnectionError) -> Self {
+        match value {
+            server_connection::ServerConnectionError::ExecutionError(message) => {
+                CliError::ExecutionError(message)
+            }
+        }
+    }
+}
